Voto personale con mezzo punto (0.5).
- Migrazione: user_media.rating da TINYINT a DECIMAL(3,1).
- Backend: set_rating arrotonda al mezzo punto (0.5-10); row_to_entry float.
- UI SeriesRating: tap sulla metà sinistra della barra = mezzo voto (n-0.5),
  metà destra = intero; barre riempibili a metà. Non invasivo, resta la fila
  di 10 barre.
- Import SeriesGuide/Cinemaniac: voti mappati a mezzo punto (non più arrotondati
  all'intero).

+++ api/lib/library.php
@@ -119,7 +119,7 @@ function library_row_to_entry(array $row, array $episodes): array {
         'id'              => $row['media_key'],
         'status'          => $row['status'],
         'favorite'        => (int) ($row['is_favorite'] ?? 0) === 1,
-        'rating'          => $row['rating'] !== null ? (int) $row['rating'] : null,
+        'rating'          => $row['rating'] !== null ? (float) $row['rating'] : null,
         'currentSeason'   => $row['current_season'] !== null ? (int) $row['current_season'] : null,
         'currentEpisode'  => $row['current_episode'] !== null ? (int) $row['current_episode'] : null,
         'watchedEpisodes' => $watched,
@@ -665,7 +665,7 @@ function library_clear_watched(PDO $pdo, string $userId, string $id, ?string $re
     return library_fetch_state($pdo, $userId);
 }
 
-function library_set_rating(PDO $pdo, string $userId, string $id, ?int $rating): array {
+function library_set_rating(PDO $pdo, string $userId, string $id, ?float $rating): array {
     $entry = library_get_entry($pdo, $userId, $id);
     $entry['rating'] = $rating;
     if (empty($entry['status'])) $entry['status'] = 'plan_to_watch';
